auth-service/app/requirements.txt
orjson==3.9.1
redis==4.6.0
elasticsearch==8.8.0
aiohttp==3.8.4
gunicorn==20.1.0
alembic==1.12.0
asyncpg==0.28.0
fastapi==0.103.1
sqlmodel==0.0.8
uvicorn==0.23.2
typer==0.9.0
pyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4
async_fastapi_jwt_auth==0.5.1
Werkzeug==2.3.6
pydantic[email]
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
python-dotenv==1.0.0
typer==0.9.0
fastapi-limiter==0.1.5
fastapi-pagination==0.12.9
python-multipart==0.0.6
